+++ week4/seonghyeon/src/components/BottomNav.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import BottomComponent from "./BottomComponent";
+import navItems from "../BottomModel";
+
+const BottomNav = () => {
+  return (
+    <div className="bottom-nav">
+      {navItems.map((navItem) => (
+        <BottomComponent navItem={navItem} key={navItem.id} />
+      ))}
+    </div>
+  );
+};
+
+export default BottomNav;
