+++ week4/seonghyeon/src/components/Content.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import Item from "./Item";
+import marketModel from "../MarketModel";
+
+const Content = () => {
+  return (
+    <div>
+      {marketModel.map((item) => (
+        <Item item={item} key={item.id} />
+      ))}
+    </div>
+  );
+};
+
+export default Content;
