+++ week4/seonghyeon/src/components/Header.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const Header = () => {
+  return (
+    <header className="header">
+      <span className="location">군자동 ▼</span>
+      <div className="icons">
+        <img src="./assets/right-chevron.svg" alt="Search" className="icon" />
+      </div>
+    </header>
+  );
+};
+
+export default Header;
