+++ week4/seonghyeon/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from "react-dom/client";
 import Home from "./Home"; // Home을 기본으로 가져오기
-import marketModel from './MarketModel';
+
 
 const root = createRoot(document.getElementById("app"));
 root.render(<Home/>); //
